requirements.txt
APScheduler==3.9.1
Flask==2.1.1
Flask_SocketIO==5.1.1
PyNaCl==1.5.0
gevent==21.12.0
gevent-websocket==0.10.1